Refactor interpreter debug hook tests
